去掉单独添加用户等级的migrate  修改商户等级字段类型

+++ database/migrations/2018_05_30_135029_alert_users_and_merchants_table_add_level_column.php
@@ -18,7 +18,7 @@ class AlertUsersAndMerchantsTableAddLevelColumn extends Migration
             $table->tinyInteger('level')->index()->default(0)->comment('用户等级 1-萌新 2-粉丝 3-铁杆 4-骨灰');
         });
         Schema::table('merchants', function (Blueprint $table) {
-            $table->integer('level')->index()->default(0)->comment('商户等级 1-签约商户 2-联盟商户 3-品牌商户');
+            $table->tinyInteger('level')->index()->default(0)->comment('商户等级 1-签约商户 2-联盟商户 3-品牌商户');
         });
     }
